# storage/docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OmPay API",
    "description": "API de paiement mobile OmPay avec authentification OTP",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Serveur de développement local"
    },
    {
      "url": "https://ompay-api-ou4t.onrender.com",
      "description": "Serveur de production Render"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/otp/verify": {
      "post": {
        "summary": "Vérifier un code OTP",
        "description": "Vérifie le code OTP et retourne un token JWT pour la connexion ou confirme l'inscription",
        "tags": ["OTP"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["telephone", "code", "type"],
                "properties": {
                  "telephone": {"type": "string", "example": "+221771234567"},
                  "code": {"type": "string", "example": "0939"},
                  "type": {"type": "string", "enum": ["inscription", "connexion", "transaction"]}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Code OTP vérifié",
            "content": {
              "application/json": {
                "examples": {
                  "inscription": {
                    "summary": "Vérification d'inscription",
                    "value": {
                      "succes": true,
                      "message": "Code OTP vérifié avec succès",
                      "donnees": {
                        "telephone": "+221771234567",
                        "type": "inscription",
                        "verifie": true,
                        "utilisateur_existe": true,
                        "utilisateur": {
                          "id": 7,
                          "uuid": "0549f1b0-2db1-4b9a-bf89-0ed594bb0bd0",
                          "nom": "Jean Dupont",
                          "telephone": "+221771234567",
                          "email": null,
                          "qr_code": "iVBORw0KGgoAAAANSUhEUgAA...",
                          "created_at": "2025-11-11T11:47:37.000000Z",
                          "updated_at": "2025-11-11T11:47:37.000000Z"
                        },
                        "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9..."
                      }
                    }
                  },
                  "connexion": {
                    "summary": "Vérification de connexion (avec token)",
                    "value": {
                      "succes": true,
                      "message": "Code OTP vérifié avec succès",
                      "donnees": {
                        "telephone": "+221771234567",
                        "type": "connexion",
                        "verifie": true,
                        "utilisateur": {
                          "id": 1,
                          "uuid": "550e8400-e29b-41d4-a716-446655440000",
                          "nom": "Jean Dupont",
                          "telephone": "+221771234567",
                          "qr_code": "iVBORw0KGgoAAAANSUhEUgAA..."
                        },
                        "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9..."
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Code OTP invalide ou expiré",
            "content": {
              "application/json": {
                "example": {
                  "succes": false,
                  "message": "Code OTP invalide ou expiré"
                }
              }
            }
          },
          "422": {
            "description": "Données de requête invalides",
            "content": {
              "application/json": {
                "example": {
                  "succes": false,
                  "message": "Les données fournies ne sont pas valides",
                  "erreurs": {
                    "code": ["Le code OTP doit contenir exactement 4 caractères"]
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne du serveur",
            "content": {
              "application/json": {
                "example": {
                  "succes": false,
                  "message": "Erreur lors de la vérification du code OTP"
                }
              }
            }
          }
        }
      }
    },
...
  }
}